Try fix DynamicButtonsTemplate on older images
Move StaticText buttons to the end of the screen. ObjectTypes as skin widget parameter is introduced later, so if buttons are caused error, then do not disturb the Button and Label buttons.

Use own converter SGConditionalShowHide. Boolean property in StaticText is introduced later, so it does not work in this image version.